Review changes from Epix. Added test for PowerHandler.
