+++ Hearthstone Deck Tracker/LogReader/Handlers/AssetHandler.cs
